apps/api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
faiss-cpu
sentence-transformers
numpy
PyPDF2
httpx
